+++ src/main/java/com/jasik/momsnaggingapi/domain/push/controller/PushController.java
@@ -1,0 +1,7 @@
+package com.jasik.momsnaggingapi.domain.push.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class PushController {
+}
