+++ src/main/java/com/jasik/momsnaggingapi/domain/push/repository/PushRepository.java
@@ -1,0 +1,7 @@
+package com.jasik.momsnaggingapi.domain.push.repository;
+
+import com.jasik.momsnaggingapi.domain.push.Push;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PushRepository extends JpaRepository<Push, Long> {
+}
